fs: change variable addr_space_sz type to uint64_t from size_t
addr_space_sz is expected to store a 64 bit value, but size_t
can vary based on the architecture it is running on, so change
addr_space_sz to explicitly to be uint64_t.

+++ fs/zbd_zenfs.cc
@@ -202,7 +202,7 @@ IOStatus ZonedBlockDevice::CheckScheduler() {
 IOStatus ZonedBlockDevice::Open(bool readonly) {
   struct zbd_zone *zone_rep;
   unsigned int reported_zones;
-  size_t addr_space_sz;
+  uint64_t addr_space_sz;
   zbd_info info;
   Status s;
   uint64_t i = 0;
